refactor : 장소와의 공유 컴포넌트를 위해 위치 이동
