+++ linkup/src/components/common/MapDisplay.vue
